A few more bug fixes and hacks for compilers. Checks for clang-3.9.0 and gcc-5.4.0 underway, they break ETs
Former-commit-id: c638420163312b59a863585d821650abd70f46da
Former-commit-id: e94ff0327022eb5cc2cf9bedf0748dd60bc0ceb5

+++ meta/einsum_meta.h
@@ -128,7 +128,7 @@ constexpr std::array<size_t,sizeof...(Rest)> nprods<Index<Rest...>,std_ext::inde
 // this is a meta-function equivalent to numpy's "where"
 template<size_t N>
 constexpr int find_index(const size_t (&ind)[N], int num, size_t i=0){
-    return (i==N) ? N : (ind[i]==num ? i : find_index(ind,num,i+1));
+    return (i==N) ? N : (static_cast<int>(ind[i])==num ? i : find_index(ind,num,i+1));
 }
 
 // check if a given value is ind1 and not ind0
